requirements.txt
streamlit
pdfplumber
google-generativeai
playwright